+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Re-throw errors during command run in Zettlr main class.
 - Moved the dictionary to its own dedicated provider for more versatility and improved upon its functionality.
 - Created an appearance provider which takes care of switching the Zettlr theming based upon user choices.
+- Switched from the `build`-property `electron-build` toolchain to the API.
 
 # 1.2.3